wip: t2127 first round of higgsfield file-complexity decomposition
